backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
numpy
pandas
python-dotenv
aiohttp
geoalchemy2
deep-translator
xgboost
scikit-learn
tensorflow
pillow
